Refactor data validation function and update imports
